+++ e2e/swaps/e2e-swap-suite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @coinmasters/e2e-transfer-dogecoin
 
+## 1.8.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.15
+
 ## 1.8.14
 
 ### Patch Changes