+++ e2e/transfers/e2e-transfer-suite/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/e2e-transfer-dogecoin
 
+## 1.8.15
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/pioneer-sdk@4.8.15
+
 ## 1.8.14
 
 ### Patch Changes